@@ -23,20 +23,30 @@
   "require": {
     "php": "^8.0",
     "ext-json": ">=1.3.7",
-    "ext-curl": "*",
-    "ezimuel/ringphp": "^1.2.2",
+    "php-http/async-client-implementation": "^1.0",
+    "php-http/discovery": "^1.20",
+    "php-http/guzzle7-adapter": "^1.0",
+    "psr/http-client": "^1.0",
+    "psr/http-client-implementation": "^1.0",
+    "psr/http-factory": "^1.1",
+    "psr/http-factory-implementation": "^2.4",
+    "psr/http-message": "^2.0",
+    "psr/http-message-implementation": "^1.0",
     "psr/log": "^1|^2|^3",
     "symfony/yaml": "*"
   },
   "require-dev": {
     "ext-zip": "*",
     "aws/aws-sdk-php": "^3.0",
-    "friendsofphp/php-cs-fixer": "^3.0",
-    "mockery/mockery": "^1.2",
-    "phpstan/phpstan": "^1.7.15",
-    "phpstan/phpstan-mockery": "^1.1.0",
-    "phpunit/phpunit": "^9.3",
-    "symfony/finder": "~4.0 || ~5.0"
+    "friendsofphp/php-cs-fixer": "^v3.64",
+    "guzzlehttp/psr7": "^2.7",
+    "mockery/mockery": "^1.6",
+    "phpstan/phpstan": "^1.12",
+    "phpstan/phpstan-mockery": "^1.1",
+    "phpunit/phpunit": "^9.6",
+    "symfony/finder": "^6.4|^7.0",
+    "symfony/http-client": "^7.1",
+    "symfony/http-client-contracts": "^3.5"
   },
   "suggest": {
     "monolog/monolog": "Allows for client-level logging and tracing",
@@ -54,7 +64,10 @@
     }
   },
   "config": {
-    "sort-packages": true
+    "sort-packages": true,
+    "allow-plugins": {
+      "php-http/discovery": true
+    }
   },
   "scripts": {
     "php-cs": [
