@@ -22,26 +22,27 @@ declare(strict_types=1);
 namespace OpenSearch\Endpoints;
 
 use OpenSearch\Common\Exceptions\UnexpectedValueException;
+use OpenSearch\EndpointInterface;
 use OpenSearch\Serializers\SerializerInterface;
 
 use function array_filter;
 
-abstract class AbstractEndpoint
+abstract class AbstractEndpoint implements EndpointInterface
 {
     /**
      * @var array
      */
-    protected $params = [];
+    protected array $params = [];
 
     /**
      * @var string|null
      */
-    protected $index = null;
+    protected ?string $index = null;
 
     /**
      * @var string|int|null
      */
-    protected $id = null;
+    protected string|int|null $id = null;
 
     /**
      * @var string|null
@@ -51,17 +52,17 @@ abstract class AbstractEndpoint
     /**
      * @var string|array|null
      */
-    protected $body = null;
+    protected string|array|null $body = null;
 
     /**
      * @var array
      */
-    private $options = [];
+    private array $options = [];
 
     /**
      * @var SerializerInterface
      */
-    protected $serializer;
+    protected SerializerInterface $serializer;
 
     /**
      * @return string[]
@@ -83,9 +84,8 @@ abstract class AbstractEndpoint
      * Set the parameters for this endpoint
      *
      * @param mixed[] $params Array of parameters
-     * @return $this
      */
-    public function setParams(array $params)
+    public function setParams(array $params): static
     {
         $this->extractOptions($params);
         $this->checkUserParams($params);
@@ -117,7 +117,7 @@ abstract class AbstractEndpoint
      *
      * @return $this
      */
-    public function setIndex($index)
+    public function setIndex($index): static
     {
         if ($index === null) {
             return $this;
@@ -134,12 +134,7 @@ abstract class AbstractEndpoint
         return $this;
     }
 
-    /**
-     * @param int|string|null $docID
-     *
-     * @return $this
-     */
-    public function setId($docID)
+    public function setId(int|string|null $docID): static
     {
         if ($docID === null) {
             return $this;
@@ -154,16 +149,12 @@ abstract class AbstractEndpoint
         return $this;
     }
 
-    /**
-     * @return array|string
-     */
-    public function getBody()
+    public function getBody(): array|string
     {
         return $this->body;
     }
 
-
-    public function setBody(array $body)
+    public function setBody(array|string $body): static
     {
         $this->body = $body;
 
