@@ -21,158 +21,85 @@ declare(strict_types=1);
 
 namespace OpenSearch;
 
-use OpenSearch\Common\Exceptions;
-use OpenSearch\ConnectionPool\AbstractConnectionPool;
-use OpenSearch\Connections\Connection;
-use OpenSearch\Connections\ConnectionInterface;
-use GuzzleHttp\Ring\Future\FutureArrayInterface;
-use Psr\Log\LoggerInterface;
+use Http\Client\HttpAsyncClient;
+use Http\Discovery\HttpAsyncClientDiscovery;
+use Http\Promise\Promise;
+use OpenSearch\Common\Exceptions\NoAsyncClientException;
+use Psr\Http\Client\ClientInterface;
+use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
 
-class Transport
+final class Transport implements TransportInterface
 {
-    /**
-     * @var AbstractConnectionPool
-     */
-    public $connectionPool;
-
-    /**
-     * @var LoggerInterface
-     */
-    private $log;
-
-    /**
-     * @var int
-     */
-    public $retryAttempts = 0;
-
-    /**
-     * @var ConnectionInterface
-     */
-    public $lastConnection;
-
-    /**
-     * @var int
-     */
-    public $retries;
+    private ?HttpAsyncClient $asyncClient = null;
 
     /**
      * Transport class is responsible for dispatching requests to the
      * underlying cluster connections
-     *
-     * @param int                                   $retries
-     * @param bool                                  $sniffOnStart
-     * @param ConnectionPool\AbstractConnectionPool $connectionPool
-     * @param \Psr\Log\LoggerInterface              $log            Monolog logger object
      */
-    public function __construct(int $retries, AbstractConnectionPool $connectionPool, LoggerInterface $log, bool $sniffOnStart = false)
-    {
-        $this->log            = $log;
-        $this->connectionPool = $connectionPool;
-        $this->retries        = $retries;
+    public function __construct(
+        protected ClientInterface $client,
+        protected RequestFactoryInterface $requestFactory,
+    ) {
+    }
 
-        if ($sniffOnStart === true) {
-            $this->log->notice('Sniff on Start.');
-            $this->connectionPool->scheduleCheck();
+    /**
+     * Create a new request.
+     */
+    public function createRequest(string $method, string $uri, array $params = [], mixed $body = null): RequestInterface
+    {
+        return $this->requestFactory->createRequest($method, $uri, $params, $body);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function sendAsyncRequest(RequestInterface $request): Promise
+    {
+        $httpAsyncClient = $this->getAsyncClient();
+        return $httpAsyncClient->sendAsyncRequest($request);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function sendRequest(RequestInterface $request): ResponseInterface
+    {
+        return $this->client->sendRequest($request);
+    }
+
+    /**
+     * Set the async client to use for async requests.
+     */
+    public function setAsyncClient(HttpAsyncClient $asyncClient): self
+    {
+        $this->asyncClient = $asyncClient;
+        return $this;
+    }
+
+    /**
+     * Get the async client to use for async requests.
+     *
+     * If no async client is set, the discovery mechanism will be used to find
+     * an async client.
+     *
+     * @throws NoAsyncClientException
+     */
+    private function getAsyncClient(): HttpAsyncClient
+    {
+        if ($this->asyncClient) {
+            return $this->asyncClient;
         }
-    }
 
-    /**
-     * Returns a single connection from the connection pool
-     * Potentially performs a sniffing step before returning
-     */
-    public function getConnection(): ConnectionInterface
-    {
-        return $this->connectionPool->nextConnection();
-    }
+        if ($this->client instanceof HttpAsyncClient) {
+            return $this->asyncClient = $this->client;
+        }
 
-    /**
-     * Perform a request to the Cluster
-     *
-     * @param string     $method  HTTP method to use
-     * @param string     $uri     HTTP URI to send request to
-     * @param array<string, mixed> $params  Optional query parameters
-     * @param mixed|null $body    Optional query body
-     * @param array      $options
-     *
-     * @throws Common\Exceptions\NoNodesAvailableException|\Exception
-     */
-    public function performRequest(string $method, string $uri, array $params = [], $body = null, array $options = []): FutureArrayInterface
-    {
         try {
-            $connection  = $this->getConnection();
-        } catch (Exceptions\NoNodesAvailableException $exception) {
-            $this->log->critical('No alive nodes found in cluster');
-            throw $exception;
+            return $this->asyncClient = HttpAsyncClientDiscovery::find();
+        } catch (\Exception $e) {
+            throw new NoAsyncClientException('No async HTTP client found. Install a package providing "php-http/async-client-implementation"', 0, $e);
         }
-
-        $response             = [];
-        $caughtException      = null;
-        $this->lastConnection = $connection;
-
-        $future = $connection->performRequest(
-            $method,
-            $uri,
-            $params,
-            $body,
-            $options,
-            $this
-        );
-
-        $future->promise()->then(
-            //onSuccess
-            function ($response) {
-                $this->retryAttempts = 0;
-                // Note, this could be a 4xx or 5xx error
-            },
-            //onFailure
-            function ($response) {
-                $code = $response->getCode();
-                // Ignore 400 level errors, as that means the server responded just fine
-                if ($code < 400 || $code >= 500) {
-                    // Otherwise schedule a check
-                    $this->connectionPool->scheduleCheck();
-                }
-            }
-        );
-
-        return $future;
     }
 
-    /**
-     * @param FutureArrayInterface $result  Response of a request (promise)
-     * @param array                $options Options for transport
-     *
-     * @return callable|array
-     */
-    public function resultOrFuture(FutureArrayInterface $result, array $options = [])
-    {
-        $response = null;
-        $async = isset($options['client']['future']) ? $options['client']['future'] : null;
-        if (is_null($async) || $async === false) {
-            do {
-                $result = $result->wait();
-            } while ($result instanceof FutureArrayInterface);
-        }
-        return $result;
-    }
-
-    public function shouldRetry(array $request): bool
-    {
-        if ($this->retryAttempts < $this->retries) {
-            $this->retryAttempts += 1;
-
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * Returns the last used connection so that it may be inspected.  Mainly
-     * for debugging/testing purposes.
-     */
-    public function getLastConnection(): ConnectionInterface
-    {
-        return $this->lastConnection;
-    }
 }
