@@ -21,10 +21,13 @@ declare(strict_types=1);
 
 namespace OpenSearch\Connections;
 
-use OpenSearch\Serializers\SerializerInterface;
 use OpenSearch\Transport;
-use Psr\Log\LoggerInterface;
 
+@trigger_error(__CLASS__ . ' is deprecated in 2.3.2 and will be removed in 3.0.0.', E_USER_DEPRECATED);
+
+/**
+ * @deprecated in 2.3.2 and will be removed in 3.0.0.
+ */
 interface ConnectionInterface
 {
     /**
